+++ UnitTests/PaymentButtonsTests/PayPalCreditButton_Tests.swift
@@ -18,6 +18,30 @@ class PayPalCreditButton_Tests: XCTestCase {
         XCTAssertNil(sut.label)
     }
 
+    func testInit_whenWhitePayPalButtonCreated_hasBorderConfiguration() {
+        let sut = PayPalCreditButton(color: .white)
+
+        XCTAssertEqual(sut.containerView.layer.borderWidth, 1.0)
+        let expectedColor = UIColor(hexString: "#555555").cgColor
+        XCTAssertEqual(sut.containerView.layer.borderColor, expectedColor)
+    }
+
+    func testInit_whenBluePayPalButtonCreated_hasNoBorderConfiguration() {
+        let sut = PayPalCreditButton(color: .blue)
+
+        XCTAssertEqual(sut.containerView.layer.borderWidth, 0.0)
+        let expectedColor = UIColor.clear.cgColor
+        XCTAssertEqual(sut.containerView.layer.borderColor, expectedColor)
+    }
+
+    func testInit_whenBlackPayPalButtonCreated_hasNoBorderConfiguration() {
+        let sut = PayPalCreditButton(color: .black)
+
+        XCTAssertEqual(sut.containerView.layer.borderWidth, 0.0)
+        let expectedColor = UIColor.clear.cgColor
+        XCTAssertEqual(sut.containerView.layer.borderColor, expectedColor)
+    }
+
     // MARK: - PayPalPayCreditButton.Representable for SwiftUI
     
     func testMakeCoordinator_whenOnActionIsCalled_executesActionPassedInInitializer() {